Refactor: centralized http factory instance creation on a extension method to avoid code repetition;
