+++ src/backend/big_core.Api/big_core.Api.csproj
@@ -12,4 +12,7 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\big_core.Common\big_core.Common.csproj" />
+  </ItemGroup>
 </Project>
